Splitting out debug
